==> NLP with Disaster Tweets/03 modeling optuna tuning.ipynb <==
from matplotlib import pyplot as plt
import math, os, re, time, random, string
import numpy as np, pandas as pd, seaborn as sns
from tqdm import tqdm
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns

from collections import defaultdict
import wordcloud

import warnings
warnings.filterwarnings('ignore')

import tensorflow as tf

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

from tensorflow.keras.layers import *
from tensorflow.keras import Input, Model
from tensorflow.keras.regularizers import l2
from tensorflow.keras.optimizers import *
from tensorflow.keras.models import Sequential
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint, ReduceLROnPlateau
import tensorflow_addons as tfa

from sklearn.metrics import *
from sklearn.preprocessing import StandardScaler
import optuna

train = pd.read_csv('train_cleaned.csv')
test = pd.read_csv('test_cleaned.csv')
#train = pd.read_csv('train_cleaned_lite.csv')
#test = pd.read_csv('test_cleaned_lite.csv')

train['text'] = train['text'].astype(str)
test['text'] = test['text'].astype(str)

labels = train['target']
print('targets shape', labels.shape)

from sklearn.model_selection import train_test_split

X_train, X_val, y_train, y_val = train_test_split(train, labels, test_size = 0.2, stratify=labels)
X_train.shape
X_val.shape
y_train.shape
y_val.shape

#### Tokenizing

for i in [X_train, X_val, test]:
    i['text'] = i['text'].astype(str)


X_train_text = [i for i in X_train['text']]
X_val_text = [i for i in X_val['text']]
test_text = [i for i in test['text']]

tokenizer = Tokenizer()

tokenizer.fit_on_texts(X_train_text)

print(len(tokenizer.word_index), 'unique tokens')

X_train_seq = tokenizer.texts_to_sequences(X_train_text)
X_val_seq = tokenizer.texts_to_sequences(X_val_text)
test_seq = tokenizer.texts_to_sequences(test_text)

X_train_data = pad_sequences(X_train_seq)
X_val_data = pad_sequences(X_val_seq, X_train_data.shape[1])
test_data = pad_sequences(test_seq, X_train_data.shape[1])

X_train_data.shape
X_val_data.shape
test_data.shape

### Meta-feature scaling

scaler = StandardScaler()
scaler.fit(X_train.iloc[:, 2:])

meta_train = scaler.transform(X_train.iloc[:, 2:])
meta_val = scaler.transform(X_val.iloc[:, 2:])
meta_test = scaler.transform(test.iloc[:, 1:])

meta_train.shape
meta_val.shape
meta_test.shape

#### Functions

def loss_acc_plot(history):
    scores = pd.DataFrame(history.history)
    scores[['loss', 'val_loss']].plot();
    scores[['accuracy', 'val_accuracy']].plot();
    #scores[['f1_score', 'val_f1_score']].plot();

answers = pd.read_csv('answer key.csv')
y_true = answers['target']

def kaggle(model, filename=''):
    y_pred = (model.predict([test_data, meta_test]) > 0.5).astype(int)
    print('\nf1 score is:', f1_score(y_true, y_pred, average='macro'))
    print(classification_report(y_true, y_pred))
    print(confusion_matrix(y_true, y_pred))
    
    if filename != '':
        submission=pd.read_csv('sample_submission.csv')
        submission['target']=y_pred
        submission.to_csv(filename+'.csv', index=False)

reduce_lr = ReduceLROnPlateau(monitor = 'val_loss', 
                              factor = 0.5, 
                              verbose = 1, 
                              patience = 4,                        
                              min_lr = 0.0001)

def model_process(model, n_epochs=20, model_name='', filename=''):
    model.summary()
    
    checkpointer = ModelCheckpoint(filepath = model_name+'.hdf5', 
                                   save_best_only=True,
                                   save_weights_only=True,
                                   monitor='val_loss')

    history = model.fit(x=[X_train_data, meta_train],
                        y=y_train,
                        epochs = n_epochs,
                        verbose=1,
                        callbacks=[checkpointer, reduce_lr],
                        validation_data = ([X_val_data, meta_val], y_val)
                       )
    
    loss_acc_plot(history)
    
    model.load_weights(model_name+'.hdf5')
    
    loss, acc = model.evaluate([X_val_data, meta_val], y_val)
    print('\nModel accuracy on validation set = ', acc)
    print('\nModel loss on validation set = ', loss)
    
    y_val_pred = (model.predict([X_val_data, meta_val]) > 0.5).astype(int)
    print('\nModel F1 on validation set = ', f1_score(y_val, y_val_pred, average='macro'))
    
    kaggle(model, filename)

embeddings_index = {}
with open('glove.6B/glove.6B.200d.txt','r', encoding="utf8") as f:
    for line in tqdm(f):
        values = line.split()
        word = values[0]
        coefs = np.asarray(values[1:], dtype='float32')
        embeddings_index[word] = coefs
f.close()

EMBEDDING_DIM = 200  
word_index = tokenizer.word_index

#initialize embedding matrix with zeros
embedding_matrix = np.zeros((len(word_index) + 1, EMBEDDING_DIM))

#add glove word encodings to our library
for word, i in tqdm(word_index.items()):
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        
        #words not found in embedding index will be all-zeros.
        embedding_matrix[i] = embedding_vector
        
print("Our embedded matrix is of dimension", embedding_matrix.shape)

embedding = Embedding(len(word_index) + 1, EMBEDDING_DIM, weights = [embedding_matrix],
                     input_length = 23, trainable = False)  

# Optuna

def objective(trial):
    LSTM_dropout = trial.suggest_discrete_uniform('LSTM_dropout', 0.0, 0.8, 0.1)
    dropout = trial.suggest_discrete_uniform('dropout', 0.0, 0.8, 0.1)
    spatial_dropout = trial.suggest_discrete_uniform('spatial_dropout', 0.0, 0.8, 0.1)
    
    lr = trial.suggest_float("lr", 1e-5, 1e-1, log=True)
    optimizer = trial.suggest_categorical("optimizer", [Adam, RMSprop])
    
    LSTM1 = trial.suggest_int('units1', 10, 100, step=10)
    LSTM2 = trial.suggest_int('units2', 10, 100, step=10)
    dense_units = trial.suggest_int('dense_units', 10, 100, step=10)
    
        
    #model
    nlp_input = Input(shape = (23,), name = 'nlp_input')
    meta_input_train = Input(shape = (7, ), name = 'meta_train')
    
    emb = embedding(nlp_input)
    emb = SpatialDropout1D(spatial_dropout)(emb)

    nlp_out = Bidirectional(LSTM(LSTM1, dropout=LSTM_dropout, return_sequences=True))(emb)  
    nlp_out = SpatialDropout1D(spatial_dropout)(nlp_out)
    
    nlp_out = Bidirectional(LSTM(LSTM2, dropout=LSTM_dropout))(emb)   
     
    #add meta data    
    x = Concatenate()([nlp_out, meta_input_train])
    
    #add second hidden layer
    x = Dropout(dropout)(x)
    x = (Dense(dense_units, 
               activation = 'relu'
              ))(x)
    
    #add output layer
    x = Dropout(dropout)(x)
    preds = Dense(1, 
                  activation='sigmoid'
                 )(x)
    
    #compile model
    model = Model(inputs=[nlp_input , meta_input_train], outputs = preds)

    model.compile(loss = 'binary_crossentropy', optimizer = optimizer(lr), metrics = ['accuracy'])
    
    reduce_lr = ReduceLROnPlateau(monitor = 'val_loss', 
                              factor = 0.5, 
                              verbose = False, 
                              patience = 4,                        
                              min_lr = 0.0001)
    
    pruning = optuna.integration.TFKerasPruningCallback(trial, 'val_accuracy')
    
    model.fit(x=[X_train_data, meta_train],
                        y=y_train,
                        epochs = 15,
                        verbose=False,
                        callbacks=[reduce_lr, pruning],
                        validation_data = ([X_val_data, meta_val], y_val),
                        shuffle=True
                       )
    y_pred = (model.predict([test_data, meta_test]) > 0.5).astype(int)
    f1 = f1_score(y_true, y_pred, average='macro')
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=50)

trial = study.best_trial
trial.params

# LSTM with GloVe

embeddings_index = {}
with open('glove.6B/glove.6B.200d.txt','r', encoding="utf8") as f:
    for line in tqdm(f):
        values = line.split()
        word = values[0]
        coefs = np.asarray(values[1:], dtype='float32')
        embeddings_index[word] = coefs
f.close()

EMBEDDING_DIM = 200  
word_index = tokenizer.word_index

#initialize embedding matrix with zeros
embedding_matrix = np.zeros((len(word_index) + 1, EMBEDDING_DIM))

#add glove word encodings to our library
for word, i in tqdm(word_index.items()):
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        
        #words not found in embedding index will be all-zeros.
        embedding_matrix[i] = embedding_vector
        
print("Our embedded matrix is of dimension", embedding_matrix.shape)

embedding = Embedding(len(word_index) + 1, EMBEDDING_DIM, weights = [embedding_matrix],
                     input_length = 23, trainable = False)  

def m5_glove(dropout_val):
            
    #define inputs
    nlp_input = Input(shape = (40,), name = 'nlp_input')
    meta_input_train = Input(shape = (7, ), name = 'meta_train')
    
    emb = embedding(nlp_input)
    emb = SpatialDropout1D(dropout_val)(emb)

    nlp_out = Bidirectional(LSTM(100, 
                                 dropout=dropout_val, 
                                 ))(emb)     
     
    #add meta data    
    x = Concatenate()([nlp_out, meta_input_train])
    
    #add second hidden layer
#    x = Dropout(dropout_val)(x)
#    x = (Dense(50, 
#               activation = LeakyReLU(alpha=0.01)
#              ))(x)
    
    #add output layer
    x = Dropout(dropout_val)(x)
    preds = Dense(1, 
                  activation='sigmoid'
                 )(x)
    
    #compile model
    model = Model(inputs=[nlp_input , meta_input_train], outputs = preds)
    model.compile(loss = 'binary_crossentropy', optimizer = Adam(0.001), metrics = ['accuracy'])
    
    return model

model5 = m5_glove(0.4)
model_process(model5, 50, '05 GloVe LSTM')

# GloVe dual LSTM

def m6_glove2(dropout_val):
            
    #define inputs
    nlp_input = Input(shape = (40,), name = 'nlp_input')
    meta_input_train = Input(shape = (7, ), name = 'meta_train')
    

    emb = embedding(nlp_input)
    emb = SpatialDropout1D(dropout_val)(emb)

    
    nlp_out = Bidirectional(LSTM(130, 
                                 dropout=dropout_val,
                                 return_sequences=True))(emb)    
    
    nlp_out = SpatialDropout1D(dropout_val)(nlp_out)
    
    nlp_out = Bidirectional(LSTM(110, 
                                 dropout=dropout_val))(emb)    
    
     
    #add meta data    
    x = Concatenate()([nlp_out, meta_input_train])
    
    #add second hidden layer
    x = Dropout(dropout_val)(x)
    x = (Dense(90, 
               activation = 'relu'
              ))(x)
    
    #add output layer
    x = Dropout(dropout_val)(x)
    preds = Dense(1, 
                  activation='sigmoid'
                 )(x)
    
    #compile model
    model = Model(inputs=[nlp_input , meta_input_train], outputs = preds)
    model.compile(loss = 'binary_crossentropy', optimizer = Adam(0.002), metrics = ['accuracy'])
    
    return model

model6 = m6_glove2(0.4)
model_process(model6, 50, '06 GloVe dual lstm', '06 GloVe dual lstm')

# model tuning

import optuna

def objective(trial):
    LSTM_dropout = trial.suggest_discrete_uniform('dropout', 0.0, 0.8, 0.1)
    dropout = trial.suggest_discrete_uniform('dropout', 0.0, 0.8, 0.1)
    lr = trial.suggest_float("lr", 1e-5, 1e-1, log=True)
    optimizer = trial.suggest_categorical("optimizer", [Adam, RMSprop])
    
    LSTM1 = trial.suggest_int('units1', 20, 200, step=10)
    LSTM2 = trial.suggest_int('units2', 20, 200, step=10)
    dense_units = trial.suggest_int('units3', 20, 200, step=10)
    
    
    
    nlp_input = Input(shape = (40,), name = 'nlp_input')
    meta_input_train = Input(shape = (7, ), name = 'meta_train')
    
    emb = embedding(nlp_input)
    emb = SpatialDropout1D(dropout)(emb)

    nlp_out = Bidirectional(LSTM(LSTM1, 
                                 dropout=LSTM_dropout,
                                 return_sequences=True))(emb)    
    
    nlp_out = SpatialDropout1D(dropout)(nlp_out)
    
    nlp_out = Bidirectional(LSTM(LSTM2, 
                                 dropout=LSTM_dropout))(emb)    
    
     
    #add meta data    
    x = Concatenate()([nlp_out, meta_input_train])
    
    #add second hidden layer
    x = Dropout(dropout)(x)
    x = (Dense(dense_units, 
               activation = 'relu'
              ))(x)
    
    #add output layer
    x = Dropout(dropout)(x)
    preds = Dense(1, 
                  activation='sigmoid'
                 )(x)
    
    #compile model
    model = Model(inputs=[nlp_input , meta_input_train], outputs = preds)

    model.compile(loss = 'binary_crossentropy', optimizer = optimizer(lr), metrics = ['accuracy'])
    
    reduce_lr = ReduceLROnPlateau(monitor = 'val_loss', 
                              factor = 0.5, 
                              verbose = False, 
                              patience = 4,                        
                              min_lr = 0.0001)
    
    model.fit(x=[X_train_data, meta_train],
                        y=y_train,
                        epochs = 30,
                        verbose=False,
                        callbacks=[reduce_lr],
                        validation_data = ([X_val_data, meta_val], y_val),
                        shuffle=True
                       )
    y_pred = (model.predict([test_data, meta_test]) > 0.5).astype(int)
    f1 = f1_score(y_true, y_pred, average='macro')
    return f1

study = optuna.create_study(direction='maximize')
study.optimize(objective, n_trials=100)

trial = study.best_trial
trial.params

trial = study.best_trial
trial.params

glove_dual = pd.read_csv('glove single lstm.csv')

f1_score(y_true, glove_dual['target'], average='weighted')
f1_score(y_true, glove_dual['target'], average='macro')
f1_score(y_true, glove_dual['target'], average='micro')
f1_score(y_true, glove_dual['target'])

df=glove_dual.drop(columns = ['prob'], axis=1)

df.to_csv('08 glove dual lstm.csv', index=False)
